Refactor Add and Edit models to inherit from a common base
