+++ Booker/Pages/Edit.cshtml
@@ -2,7 +2,7 @@
 @model Booker.Pages.EditModel
 @{
     ViewData["Title"] = "Edytuj ogłoszenie";
-    const string include = "#add-form :not([name = '__RequestVerificationToken'])";
+    const string include = "#edit-form :not([name = '__RequestVerificationToken'])";
 }
 
 <article>
